+++ src/main/java/me/villagerunknown/coinstacks/feature/CoinStackBlocksFeatureLoader.java
@@ -41,26 +41,48 @@ public class CoinStackBlocksFeatureLoader {
 				CoinStackBlocks.LARGE_IRON_COIN_STACK,
 				CoinStackBlocks.LARGE_GOLD_COIN_STACK,
 				CoinStackBlocks.LARGE_EMERALD_COIN_STACK,
-				CoinStackBlocks.LARGE_NETHERITE_COIN_STACK
+				CoinStackBlocks.LARGE_NETHERITE_COIN_STACK,
+				CoinStackBlocks.COPPER_COIN_STACK_SLAB,
+				CoinStackBlocks.IRON_COIN_STACK_SLAB,
+				CoinStackBlocks.GOLD_COIN_STACK_SLAB,
+				CoinStackBlocks.EMERALD_COIN_STACK_SLAB,
+				CoinStackBlocks.NETHERITE_COIN_STACK_SLAB,
+				CoinStackBlocks.COPPER_COIN_STACK_BLOCK,
+				CoinStackBlocks.IRON_COIN_STACK_BLOCK,
+				CoinStackBlocks.GOLD_COIN_STACK_BLOCK,
+				CoinStackBlocks.EMERALD_COIN_STACK_BLOCK,
+				CoinStackBlocks.NETHERITE_COIN_STACK_BLOCK
 		);
 		
 		return CoinStackBlocksFeature.registerCoinStackBlockEntities( builder );
 	}
 	
-	public static Block registerCoinStackBlock( CoinType type, String id, int value ) {
-		return CoinStackBlocksFeature.registerCoinStackBlock( type, id, new CoinStackBlock( AbstractBlock.Settings.create() ), value );
+	public static Block registerCoinStackBlock( String id, Block block, int value ) {
+		return CoinStackBlocksFeature.registerCoinStackBlock( id, block, value );
+	}
+	
+	public static Block registerCraftableCoinStackBlock( CoinType type, String id, Block block, int value ) {
+		return CoinStackBlocksFeature.registerCraftableCoinStackBlock( type, id, block, value );
 	}
 	
 	public static Block registerSmallCoinStackBlock( CoinType type, String id, int value ) {
-		return CoinStackBlocksFeature.registerCoinStackBlock( type, id, new SmallCoinStackBlock( AbstractBlock.Settings.create() ), value );
+		return registerCraftableCoinStackBlock( type, id, new SmallCoinStackBlock( AbstractBlock.Settings.create() ), value );
 	}
 	
 	public static Block registerMediumCoinStackBlock( CoinType type, String id, int value ) {
-		return CoinStackBlocksFeature.registerCoinStackBlock( type, id, new MediumCoinStackBlock( AbstractBlock.Settings.create() ), value );
+		return registerCraftableCoinStackBlock( type, id, new MediumCoinStackBlock( AbstractBlock.Settings.create() ), value );
 	}
 	
 	public static Block registerLargeCoinStackBlock( CoinType type, String id, int value ) {
-		return CoinStackBlocksFeature.registerCoinStackBlock( type, id, new LargeCoinStackBlock( AbstractBlock.Settings.create() ), value );
+		return registerCraftableCoinStackBlock( type, id, new LargeCoinStackBlock( AbstractBlock.Settings.create() ), value );
+	}
+	
+	public static Block registerCoinStackSlabBlock( CoinType type, String id, int value ) {
+		return registerCraftableCoinStackBlock( type, id, new CoinStackSlabBlock( AbstractBlock.Settings.create() ), value );
+	}
+	
+	public static Block registerCoinStackBlock( CoinType type, String id, int value ) {
+		return registerCraftableCoinStackBlock( type, id, new CoinStackBlock( AbstractBlock.Settings.create() ), value );
 	}
 	
 }