+++ CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - Added recipes for Coin Stacks allowing them to be crafted in Survival mode.
 
+### Changed
+
+- Changed Emerald and Netherite Coin Stack values to be accurate to their conversions. 
+They originally had a lower value for mapmakers but have been adjusted for crafting.
+
 ## [1.0.0]
 
 _Initial release for Minecraft 1.21.1_